% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality.default}
\alias{centrality.default}
\title{Default method for centrality generic function}
\usage{
\method{centrality}{default}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{Ignored}
}
\description{
Default method for centrality generic function
}
