% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality.numeric}
\alias{centrality.numeric}
\title{Function to generate statistical centrality metrics for numeric class object}
\usage{
\method{centrality}{numeric}(x, remove_na = TRUE)
}
\arguments{
\item{x}{Numeric vector}

\item{remove_na}{a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
List object that summarize numeric vector into mean, median and mode
}
\description{
Function to generate statistical centrality metrics for numeric class object
}
