% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{S3 Generic function for statistic centrality}
\usage{
centrality(x, ...)
}
\arguments{
\item{x}{data.frame or vector type object class}

\item{...}{Additional arguments affecting the centrality metrics produced.}
}
\description{
S3 Generic function for statistic centrality
}
\examples{
\dontrun{

#Centrality for data.frame object

centrality_iris <- centrality(iris)


#Centrality for numeric object

centrality_num <- centrality(c(1, 2, 3, 3, 4, 4, 4))

}

}
