% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution.default}
\alias{distribution.default}
\title{Default method for distribution generic function}
\usage{
\method{distribution}{default}(x)
}
\arguments{
\item{x}{Object}
}
\description{
Default method for distribution generic function
}
