% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{S3 Generic function for statistical distribution or quantiles values}
\usage{
distribution(x, ...)
}
\arguments{
\item{x}{data.frame or vector type object class}

\item{...}{Additional arguments affecting the distribution values produced.
e.g: \code{remove_na}: a logical evaluating to TRUE or FALSE indicating whether NA values should be stripped before the computation proceeds.
\code{probs}: numeric vector of probabilities with values in value from 0 to 1 (e.g: default c(0, 0.25, 0.5, 0.75, 1))}
}
\value{
A list of statistical distribution which include min, max, and quantile values of numeric values from data.frame or vector object
}
\description{
S3 Generic function for statistical distribution or quantiles values
}
\examples{
\dontrun{

#Distribution for numeric columns from data.frame
distribution_iris <- distribution(iris, remove_na = TRUE)



#Distribution for numeric's vector
#remove_na is default to TRUE. Missing values and NaN's not allowed if 'remove_na' is FALSE

distribution_num <- distribution(c(1, 2, 3, 3, 4, 4, 4, NA), remove_na = TRUE)

}


}
