% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreme-package.R
\docType{package}
\name{exploreme-package}
\alias{exploreme}
\alias{exploreme-package}
\title{exploreme: Provide Statistical and Analytic Summary for Dataset}
\description{
YourPackageName is an R package designed to empower data analysts, scientists, and researchers with a comprehensive toolkit for gaining deep insights from their data. Harness the power of advanced statistical and analytic techniques to uncover hidden patterns, trends, and valuable knowledge in your datasets.
}
\author{
\strong{Maintainer}: Mohd Redzuan Abdul Rahman \email{redzu.m@gmail.com}

}
\keyword{internal}
