% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centrality.R
\name{centrality}
\alias{centrality}
\title{S3 Generic function for statistic centrality}
\usage{
centrality(x, ...)
}
\arguments{
\item{x}{Support data.frame and vector object class}
}
\description{
S3 Generic function for statistic centrality
}
